fix: surfpool test reliability
